+++ packages/core/src/node/dev.ts
@@ -1,6 +1,5 @@
-import { UserConfig, removeLeadingSlash } from '@rspress/shared';
-import type { RsbuildConfig } from '@rsbuild/core';
-import { logger } from '@rspress/shared/logger';
+import { UserConfig } from '@rspress/shared';
+import { type RsbuildConfig } from '@rsbuild/core';
 import { initRsbuild } from './initRsbuild';
 import { writeSearchIndex } from './searchIndex';
 import { PluginDriver } from './PluginDriver';
@@ -18,7 +17,6 @@ interface DevOptions {
 
 export async function dev(options: DevOptions): Promise<ServerInstance> {
   const { docDirectory, config, extraBuilderConfig } = options;
-  const base = config?.base ?? '';
   const isProd = false;
   const pluginDriver = new PluginDriver(config, isProd);
   await pluginDriver.init();
@@ -26,6 +24,7 @@ export async function dev(options: DevOptions): Promise<ServerInstance> {
   try {
     const modifiedConfig = await pluginDriver.modifyConfig();
     await pluginDriver.beforeBuild();
+
     // empty temp dir before build
     // await fs.emptyDir( TEMP_DIR);
     const builder = await initRsbuild(
@@ -35,19 +34,14 @@ export async function dev(options: DevOptions): Promise<ServerInstance> {
       false,
       extraBuilderConfig,
     );
+    builder.onDevCompileDone(async () => {
+      await pluginDriver.afterBuild();
+    });
     const { server } = await builder.startDevServer({
-      printURLs: urls => {
-        return urls.map(({ label, url }) => ({
-          label,
-          url: `${url}/${removeLeadingSlash(base)}`,
-        }));
-      },
       // We will support the following options in the future
       getPortSilently: true,
-      logger,
     });
 
-    await pluginDriver.afterBuild();
     return server;
   } finally {
     await writeSearchIndex(config);