+++ packages/plugin-shiki/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @rspress/plugin-shiki
 
+## 1.10.1
+
+### Patch Changes
+
+- @rspress/shared@1.10.1
+
+## 1.10.0
+
+### Patch Changes
+
+- 0d150eb: types: export PluginShikiOptions type
+- Updated dependencies [cfee5d2]
+- Updated dependencies [c55b967]
+  - @rspress/shared@1.10.0
+
 ## 1.9.3
 
 ### Patch Changes